+++ FitnessViewer.Infrastructure/Models/Dto/RunningTimes.cs
@@ -12,7 +12,7 @@ namespace FitnessViewer.Infrastructure.Models.Dto
             public string ActivityName { get; set; }
             public DateTime ActivityDate { get; set; }
             public string DistanceName { get; set; }
-            public float Distance { get; set; }
+            public decimal Distance { get; set; }
             public TimeSpan Time { get; set; }
             public long ActivityId { get; set; }
